spurious_dragon refactored
